requirements.txt
# API clients
anthropic
openai
google-generativeai

# Training & Inference
torch>=2.6
transformers>=4.51.0
trl
peft
datasets
accelerate
bitsandbytes
wandb

# Utilities
tqdm
